Bring back a KCM for the screen locker
Together with adding back the KCM the settings are slightly adjusted to
have them fit the UI without needing a transformation in the KCM.

* timeout is stored in minutes (used to be seconds)
* grace time is stored in seconds (used to be msec)
* lock is enabled by default

Also disabling the locking after idle timeout is no longer possible. An
endless grace made some sense with the screensavers, but not with the
screenlocker. In order to not break an unknown but maybe valid use case
the option is still available, just not exposed through UI.

REVIEW: 118038