[screenlocker] Support OnScreenDisplay messages
KSld connects to the DBus signals for the OnScreenDisplay and forwards
them to the greeter application through the private Wayland protocol.
The idea here is to not have the greeter connect to the DBus as in future
it would be preferable if the greeter cannot connect to DBus.

The lookandfeel package might expose an object called "onScreenDisplay".
Whenever an OSD message is recieved by the greeter it looks for such an
object and updates the osd specific properties.

The default breeze lookandfeel package integrates the OsdItem from the
same lookandfeel package.

REVIEW: 122452