[screenlocker] Start grace time when the greeter is shown
So far the grace time started with the idle timeout. Given that our
greeter takes a few seconds to start it rendered a short grace time
useless as it was quite likely that the grace time already ended when
the first visible feedback was shown to the user.

Now the grace time starts when the greeter is actually shown.

REVIEW: 124917