[screenlocker] More XFlush and XSync before server grab
We seem to have a problem in the autotest concerning grabbing so
try to make the whole thing less timing dependent.

The XSync should be before the server grab, otherwise it just doesn't
make sense. And all XLib code should do a flush as otherwise we don't
know if that will happen at all.