Use PATH_VARS feature of ecm_configure_package_config_file
This makes specifying absolute paths on the commandline that are then
encoded into the config file work.
Not hard-coding the packackage prefix is also helpful on a multiarch
layout where the prefix is /usr/${host} but arch-independent files
should still be installed to /usr/share (i.e a level below the
prefix).

REVIEW: 125844