Disable ptrace for kcheckpass and the greeter
Setting the PR_SET_DUMPABLE flag to 0 for the security relevant
command kcheckpass and kscreenlocker_greet. If one wants to gdb into
the running command it will result in:

ptrace: Operation not permitted.

For kscreenlocker_greet ptrace is permitted in testing mode.

As root it's still possible to attach to the process.

REVIEW: 126203