Improve setting global lock shortcut
Summary:
KSld is a library which means the componentName is set to the application
name. This works fine for KSMServer, but in KWin the shortcut didn't
work.

This change manually sets the proper component name (ksmserver) on the
QAction and at the same time removes the usage of the KActionCollection.

Due to that we can also drop the dependency on XmlGui.

Test Plan: Locking screen through configured shortcut works now in kwin_wayland.

Reviewers: bshah

Subscribers: plasma-devel

Differential Revision: https://phabricator.kde.org/D820