GlobalAccel can operate on QKeyEvent
So far the GlobalAccel component was bound to X11 by only processing
native xcb events. The evaluation of whether a shortcut was triggered
is not bound to X11, though. It completely operates on Qt key codes and
modifiers. Information also available in QKeyEvent.

This change makes GlobalAccel perform the same way on QKeyEvent as on
xcb events. To achieve this KSldApp implements the virtual method event
and passes all KeyPress events sent to KSldApp to GlobalAccel. This way
a Wayland compositor (e.g. KWin) can send a QKeyEvent to KSldApp to get
the GlobalAccel support working.

Reviewed-By: Bhushan Shah, Kai-Uwe Broulik