Add autotest to verify that screen starts to lock on idle timeout
Summary:
This adds a new auto test to verify that gets locked when the idle
timeout is reached.

Unfortunately we cannot just modify the configuration as the minimum
idle timout possible through the configuration is one minute and we
don't want to wait that long on the CI system.

Thus two new methods are exposed to modify the internal idle id from
the auto test.

Test Plan: Tested with an Xvfb

Reviewers: #plasma

Subscribers: plasma-devel

Projects: #plasma

Differential Revision: https://phabricator.kde.org/D1311