Workaround problems with Qt::QueuedConnection
Summary:
For unknown reasons our signals with Qt::QueuedConnection are not
delivered with Qt 5.5.1 or later if the context object is this
(KSldApp instance).

If we use a different context object (e.g. the sender) it works.

The lack of signals not working triggered quite a few bugs, like
* grace time not working
* global shortcuts not working

BUG: 361008
BUG: 361007
FIXED-IN: 5.6.2

Reviewers: #plasma

Subscribers: plasma-devel

Projects: #plasma

Differential Revision: https://phabricator.kde.org/D1314