Load a Wallpaper plugin in kscreenlocker_greet
Summary:
With this change the greeter loads a configured Wallpaper Plugin. The
idea behind this is to bring back the possibility of "animation" into
the lockscreen and to replace the rather hacky solution for a custom
background image.

As this is security relevant code we need to consider the security
implications of such a change. The existing code already allows to load
different QML files through the lookandfeel infrastructure. Given that
local data can be loaded anyway and network data is blocked.

For a proper integration there are further changes needed:
* default lnf package needs to "remove" the rendered image to be able
  to see the wallpaper
* blurred lockscreen background by default
* config transition for custom wallpaper images

To test manually edit kscreenlockerrc and add the following:
[Greeter]
WallpaperPlugin=org.kde.hunyango

And modify the lookandfeel package to be an Item instead of an Image.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D1800