[GreeterApp] Set wallpaperGraphicsObject to what the code actually expects
When setting the view geometries, the code tries to ensure the wallpaper is loaded at the correct
size already by completing initialization of the view with correct arguments. This codepath is
never executed, however, as the property contains the rootObject instead of the actually expected
QmlObjectSharedEngine.

Moreover, accessing rootObject() with initializationDelayed will force completion right away.

This patch fixes the flickering of the wallpaper plugin which reloads the because the view
size changed after it has loaded an image.

Differential Revision: https://phabricator.kde.org/D2557