[Screen Locker KCM] Put Wallpaper in a separate tab
This cleans up the UI by moving the wallpaper configuration to a separate tab
and also gives it more room.

It uses a documentMode TabWidget (no frame) as it has a lighter color and
I couldn't find an equivalent in QPalette to make QQuickWidet blend in there.

Also take into account QtQuick Layouts hardcoded default spacing of 5px
to make ComboBoxes aligned.

Differential Revision: https://phabricator.kde.org/D2753