[greeter] Clear clipboard on show
Summary:
As we have a feature to show the password it would be possible for an
unauthorized user to get access to the clipboard.

To prevent this, the clipboard gets cleared when we have the window
shown. This is late enough on Wayland to be able to clear the clipboard.

BUG: 376526
FIXED-IN: 5.10

Test Plan:
On Wayland, paste is no longer possible. X11 is untested,
there klipper might restore clipboard content.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4650