[kcheckpass] Drop the ACCEPT_ENV compatibility define
Summary:
Given the code comment this was just for compatibility to accept some
values from the environment. The env variables KCHECKPASS_USER and
KDE_PAM_ACTION are according to lxr.kde.org nowhere in KDE's source
repository defined.

I think it's time for this code to rest in peace.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4754