[kcheckpass] Drop the caller command line option
Summary:
This was used to the specify calling application, effectively the
PAM service basename. As the greeter doesn't set it, it is unused
and unknown whether that would work at all.

By dropping the option we can significantly reduce the ifdef mess
by having the same argument count for all auth backends again.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4773