Add a small test application for kcheckpass
Summary:
This is just a small tool to help development on kcheckpass.
A minimal qtquick gui which uses the Authenticator to test whether
the authentication of provided password fails/succeeded.

It uses the kcheckpass in the build directory. Thus it is not
required to install kcheckpass to perform tests. This was a problem
during development as it allows to lock unself out if kcheckpass breaks
and the screen locks.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4785