[greeter] Send the auth result through the server instead return value
Summary:
Given that we have the protocol and don't use the legacy conv any more
there is no need to go through exit code mapping. By not using exit code
we can start reusing one kcheckpass for multiple auth request and turn it
into a kind of daemon, which is a requirement for enabling seccomp in
kscreenlocker_greet.

Test Plan: Tested with the new test helper

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4806