Add support for emergency message show on Wayland
Summary:
When playing with the lockscreen I noticed that the emergency show does
not work on Wayland and one only gets a black screen if the system fails.

This change adds a property to the QWindow (in ksld), so that KWin can
read this and now that this is a window to be shown on the lock screen.
In addition the geometry of the background window gets adjusted correctly
on Wayland.

The change needs a small part also in KWin.

Reviewers: #kwin, #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5157