Fix removal of lock window on unmap
Unfortunately our tests did not run on build.kde.org and now that they
run they show breakage.

The unmap notifiy handler removes windows from the list of lock windows.
But it used the event window instead of the window. The event window is
the rootWindow which is of course no lock window. Due to this the
unmapped lock window never got removed from the list of lock windows
which could result in the restacking not working as we tried to restack
to a non-existing window resulting in an xcb error.

Normally this should go through review, but I'm skipping it as we have
a failing auto test for the issue and it's really obvious that event
window is wrong in this case. We are in a branch which compares the
event window is the rootWindow, so it really, really is just totally
wrong.

With the change the test passes again.