Fix greeter crashing with software rendering on wayland
Summary:
kscreenlocker_greet crashes inside Mesa when running with llvmpipe or softpipe.
This is because it creates file-backed buffers for DRI2 surfaces.
So just allow creation of files for the affected targets.

Test Plan:
Ran a windowed kwin_wayland session and started kscreenlocker_greet with
LIBGL_ALWAYS_SOFTWARE=1. Now it works.

Reviewers: #plasma, graesslin

Reviewed By: #plasma, graesslin

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9496