Add touchpad shortcuts to the whitelist
Summary: My laptop's ACPI has a weird habit of sending the touchpad toggle key stroke when closing the lid, so the shortcut to enable the touchpad turns out to be handy even when the screen is locked.

Test Plan: The touchpad should be enabled/disabled after pressing the button when the screen is locked.

Reviewers: #plasma, graesslin

Reviewed By: #plasma, graesslin

Subscribers: graesslin, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10066