Prevent paste in screen locker
Summary:
KScreenlocker tries to clear the clipboard on load. However, klipper
also (by default) automatically keeps the last relevant item in the
clipboard. Whilst both parts independently work correctly, Plasma as a
whole does not.

This became a problem when we added the reveal password button as it is
a data leak.

Instead of clearing the clipboard this patch replaces it with a real
entry, but with a dummy mime value that is of no value to anyone,
especially a textfield.

BUG: 388049

Test Plan:
Reproduced issue
Made this patch
Tried pasting in session
Nothing happned
Logged in
Could paste

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: mart, graesslin, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D14924