Fix fallback to built-in ui
Summary:
If the lock screen package is not available m_mainQmlPath is an empty
URL. Thus view->setSource won't do anything and the status won't change
to Error. Due to that the fallback was not activated when the lnf
package is missing.

This change also checks whether said variable is empty.

Test Plan: run kscreenlocker_greet in testing mode with manipulated XDG_DATA_DIRS

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D16945