[KCM] Improve text for "lock after wake" feature
Summary:
The word "resume" is a technical term that's only clear when suspending is mentioned
in conjunction with it. But we don't use the term "suspend" anymore; we say "sleep."
So its inverse is "wake from sleep".

This patch improves the text according to the above reasoning.

Test Plan: {F7160676, size=full}

Reviewers: #vdg, GB_2

Reviewed By: #vdg, GB_2

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22934