[KCM] Draw full QTabWidget frame
Summary:
Right now we're not drawing the full frame in the Screen Locking KCM to work around
https://bugs.kde.org/show_bug.cgi?id=410943. But this only exchanges one aesthetic
problem for another, and as such, is not an appropriate workaround.

This patch draws the full frame to fix the bug listed below,  and then we can concentrate
on fixing https://bugs.kde.org/show_bug.cgi?id=410943 so it works for everyone with
no hacks or workarounds required.

BUG: 410923
FIXED-IN: 5.17.0

Test Plan:
{F7248765}
{F7248766}

Reviewers: #vdg, #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23184