Use LayerShellQt to implement on Wayland
Using the PlasmaShell protocol has all sorts of problems because
placement is decided by the client at a position level rather than by
the application.
This adopts the layer shell and enforces the surface to be full screen
on the correct output and the compositor should be responsible for
placing it correctly.