Fix valgrind warning on startup
Before KSldApp's constructor we call QObject::QObject. This calls
setParent which triggers a ParentChangedEvent on the virtual from the
constructor (boo!)

At this point our constructor hasn't run, so m_globalAccel isn't
initiliased.

Arguably this is a false positive as it always evaluates to the same
thing, but we don't want to be reading uninitialised memory.