Move DBus logic back into the greeter
Some calls to DBus were proxied by the host (ksmserver/kwin_wayland) and
then sent via a custom socket to the greeter.

The original goal was for security and locking down the greeter, but the
DBus bus was never filtered on the greeter, so this never achieved
anything in practice. We also have DBus used in the greeter for MPRIS
and mobile's new notification proxy, so this isn't something we could
add.

This simplifies code heavily. It reduces the workload on the sensitive
app that can't crash and puts more responsibility on the app that can
(the greeter).

Long term this unlocks two new tasks:

We can re-evaluate if wayland is really the right tool to send a a few
ints down a socket only when on X11.

We can drop can_suspend context properties and use the kworkspace import
in the greeter. Which will be a nicer API for all.