Don't recommend current session
I've noticed that KScreenLocker will always recommend using virtual terminal 2 to recover from a crash, even if you are on virtual terminal 2. This MR should fix that.

Please note that I have not actually tested this for two reasons: (1) I have no idea how to make KScreenLocker crash, and even if I did, (2) I don't have a Plasma dev environment set up to crash it in (although I'd be happy to do so).