Disable logind integration for kwin's unit tests
Using real logind causes issues with CI as it's dependent on external
factors and it can leave the host machine is an unclean state.

We can't mock it like we do in the internal tests as easily and it
wouldn't be very useful to do so. We can't disable completly as
lockscreentest relies on externally emitting a signal on this object.