Wait for screenlocker UI to call close
This is important as it allows the screenlocker to show a prompt for
passwordless users before unlocking. For example mobile having a simple
swipe to unlock.

It also potentially allows a future successful login animation.

An additional guard is added so rogue QML can't trivially bypass the
user password; though pragamtically they could before if they emit the
right signal in the authenticator.