Handle fallback packages when checking for screenlocker API version
The current lookandfeel package may not contain the
"X-Plasma-ApiVersion" metadata key but at the same time also not define
a custom lockscreen, instead relying on the fallback mechanism. In that
case, we would end up setting the package path to an empty string,
causing everything else that is using the package to treat it as
invalid.

To fix that, we need to check if the lockscreen comes from the current
package and if not, check the API version of the fallback package
instead. Additionally, we also should explicitly set the package path to
the default of Breeze instead of using an empty string as path, which
implies an invalid package rather than "use the default".