WaylandLocker: Show/hide the background when requested
Currently, the background is only shown if there's a problem with the
greeter. However, in certain cases we can end up with some desktop
contents shown when locking and going to sleep. To completely avoid
this, we can ensure the lock window shows the background before we lock.
The greeter already fades from a black background to the actual screen,
so this doesn't seem to regress anything there. As a small side bonus,
it also improves the perceived responsiveness as when requesting a lock
or sleep, the black background shows up immediately while the system is
going to sleep.

CCBUG: 316734