wallpapers/image: allow wallpaper plugin to force enable the apply button
Currently there is no way to enable the apply button if the wallpaper
config is not changed, but the image wallpaper plugin needs to force
enable the apply button after an image is removed temporarily. This
allows to force enable the apply button by adding a boolean to kcm, to
let `isSaveNeeded` depend on the boolean, so when the boolean is true,
the apply button can be enabled without the need to have any change in
wallpaper config.

BUG: 438585
FIXED-IN: 5.26