Only inhibit screen locker on "ChangeScreenSettings"
This is the policy for e.g. dimming and turning off the screen,
i.e. "user can't see stuff anymore" which includes locking the screen.

"InterruptSession" is for suspend where the computer will not
continue its current tasks. The screen can still lock when I am
copying files or burning a CD.

BUG: 464119