Make the PAM service configurable and update documentation
In KDE4 it was possible to set the PAM service name to be used
for screen unlocking via a compile definition.  According to
README.pam this should be possible in KF5, but the variable names
there are obsolete.  Even if correctly interpreted, the code does
not actually take account of any such definitions and always uses
the name "kde".  Downstream users which attempt to set the
service name to correspond to the installed PAM service will find
that the screen will not unlock.

Update the documentation to remove references to KDM and kscreensaver,
and update from KDE4.  The PAM service can be configured with
KDE_COMMON_PAM_SERVICE (formerly KDE4_COMMON_PAM_SERVICE), which
sets the service name intended to be used for all PAM-aware uses -
even though there may be no other uses of PAM within Frameworks,
this is kept for compatibility.  The specific PAM service to be used
for KScreenLocker can be configured with KSCREENLOCKER_PAM_SERVICE
(formerly KSCREENSAVER_PAM_SERVICE).  The default if none is
specified remains as "kde".

Build kscreenlocker_greet taking account of the configured service,
and report what is being used for debugging.