greeter: track various properties persistently
There is a fundamental race condition between PamAuthenticator and the
greeter views. PA is a singleton that emits signals that result in
stateful changes to the views (e.g. showing the prompt message).
Meanwhile views are added when they get connected. This leads to
circumstances where pam may have sent a prompt before a view was
created, meaning the view is in an incorrect state.

To mitigate the problem we'll force-emit a number of stateful signals
whenever they are connected to. Thus ensuring that greeter views
definitely receive the most recent states.
The more long-term solution should likely be to move to property
bindings instead, so these stateful signals are now also exposed as
stateful properties.

BUG: 456210