Don't enable wayland reconnections on screenlocker greeter
The screenlock greeter is managed by the KSLDApp, which manages when we
unlock the session through use of the exit code. If kwin restarts then
then the KSLDApp is gone too, so it's not useful.

When kwin restarts if we are locked it will spawn new greeters, the
locker surviving will give us two.