Fix and modernize fallback theme
The fallback theme was using PlasmaComponents 2 and wasn't working at
all in Plasma 6.

This commits ports it to modern components so that it works again, and
gets rid of some unnecessary old code and old QML paradigms in the
process.

BUG: 467056
FIXED-IN: 6.0

To test, break your lock screen theme and lock the screen. You should see the fallback theme now:

| Greeter | Session switcher |
| ------- | ---------------- |
| ![greeter](/uploads/6bc3471eb93b63f3ced68bd1369ddec3/greeter.jpg) | ![session_switcher](/uploads/6d00d2e413f01073b2854b2d3f99fa1b/session_switcher.jpg) |

@davidedmundson @mart @davidre